Add ManageIQ::Util::FtpLib
A library for shared Net::FTP functions that can be used throughout the
ManageIQ project.  Extracted from FileDepotFtp.
